+++ src/VerticalSliceArchitectureTemplate/Features/Todos/Domain/TodoCreatedEvent.cs
@@ -1,5 +1,5 @@
 ﻿using VerticalSliceArchitectureTemplate.Common.Domain.Base.Interfaces;
 
-namespace VerticalSliceArchitectureTemplate.Common.Domain.Todos;
+namespace VerticalSliceArchitectureTemplate.Features.Todos.Domain;
 
 public record TodoCreatedEvent(TodoId TodoId) : IDomainEvent;
