+++ src/VerticalSliceArchitectureTemplate/Features/Todos/Persistence/TodoDbContext.cs
@@ -1,4 +1,4 @@
-﻿using VerticalSliceArchitectureTemplate.Common.Domain.Todos;
+﻿using VerticalSliceArchitectureTemplate.Features.Todos.Domain;
 
 // Disable as we want the partial class to be in the same namespace as the original class
 // ReSharper disable once CheckNamespace
