+++ src/VerticalSliceArchitectureTemplate/Features/Todos/Queries/GetAllTodos.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
-using VerticalSliceArchitectureTemplate.Common.Domain.Todos;
 using VerticalSliceArchitectureTemplate.Common.Extensions;
+using VerticalSliceArchitectureTemplate.Features.Todos.Domain;
 
 namespace VerticalSliceArchitectureTemplate.Features.Todos.Queries;
